+++ server/src/main/java/cash/xcl/server/exch/ExchangePostBlockChainProcessor.java
@@ -10,13 +10,11 @@ import cash.xcl.api.AllMessages;
 import cash.xcl.api.dto.CommandFailedEvent;
 import cash.xcl.api.dto.CurrentBalanceQuery;
 import cash.xcl.api.dto.CurrentBalanceResponse;
-import cash.xcl.api.dto.DepositValueCommand;
-import cash.xcl.api.dto.DepositValueEvent;
 import cash.xcl.api.dto.TransferValueEvent;
-import cash.xcl.api.dto.WithdrawValueCommand;
-import cash.xcl.api.dto.WithdrawValueEvent;
 import cash.xcl.api.exch.CancelOrderCommand;
 import cash.xcl.api.exch.CurrencyPair;
+import cash.xcl.api.exch.DepositValueCommand;
+import cash.xcl.api.exch.DepositValueEvent;
 import cash.xcl.api.exch.ExchangeCommands;
 import cash.xcl.api.exch.ExecutionReport;
 import cash.xcl.api.exch.ExecutionReportEvent;
@@ -24,6 +22,8 @@ import cash.xcl.api.exch.NewLimitOrderCommand;
 import cash.xcl.api.exch.OrderClosedEvent;
 import cash.xcl.api.exch.TransferFromExchangeCommand;
 import cash.xcl.api.exch.TransferToExchangeCommand;
+import cash.xcl.api.exch.WithdrawValueCommand;
+import cash.xcl.api.exch.WithdrawValueEvent;
 import cash.xcl.server.LocalPostBlockChainProcessor;
 import net.openhft.chronicle.core.time.TimeProvider;
