+++ server/src/main/java/cash/xcl/server/mock/MockServer.java
@@ -11,7 +11,6 @@ import cash.xcl.api.dto.CreateNewAddressCommand;
 import cash.xcl.api.dto.CreateNewAddressEvent;
 import cash.xcl.api.dto.CurrentBalanceQuery;
 import cash.xcl.api.dto.CurrentBalanceResponse;
-import cash.xcl.api.dto.DepositValueCommand;
 import cash.xcl.api.dto.EndOfRoundBlockEvent;
 import cash.xcl.api.dto.ExchangeRateEvent;
 import cash.xcl.api.dto.FeesEvent;
@@ -21,11 +20,12 @@ import cash.xcl.api.dto.TransactionBlockEvent;
 import cash.xcl.api.dto.TransactionBlockGossipEvent;
 import cash.xcl.api.dto.TransactionBlockVoteEvent;
 import cash.xcl.api.dto.TransferValueCommand;
-import cash.xcl.api.dto.WithdrawValueCommand;
 import cash.xcl.api.exch.CancelOrderCommand;
+import cash.xcl.api.exch.DepositValueCommand;
 import cash.xcl.api.exch.NewLimitOrderCommand;
 import cash.xcl.api.exch.TransferFromExchangeCommand;
 import cash.xcl.api.exch.TransferToExchangeCommand;
+import cash.xcl.api.exch.WithdrawValueCommand;
 import cash.xcl.api.util.XCLBase32;
 import cash.xcl.server.AddressService;
 import net.openhft.chronicle.bytes.Bytes;
